(PE-12299) Prepare config files before preparing platforms
Before this commit on solaris 11 the configuration file might be backed
up before the files have had their deprecated settings removed. This
meant that when the configuration files were moved back into place which
can cause puppet to fail to start on subsequent runs.

This commit ensures that config files are cleaned up before preparing
the platforms for installing puppet-agent.