(PE-12299) Add removed directories so agent run is successful
When removing the packages the /var/opt/lib dir is removed from the
system much like /etc/puppetlabs is removed from the system. The current
agent run is unable to finish successfully since it cannot save a report
to send to puppetdb and fails out after running all of the other
resources.

This commit ensures that those paths exist so that a final report can be
saved and pushed to puppetdb.