(PE-13179) remove lower version requirement in metadata.json
Prior to this change there was a lower bound on what versions of
puppet could be upgraded by the agent module. This has caused
users to rely on supplemental tools in some cases to facilitate
upgrades. In order to encourage upgrades and make the process
easier, we're removing the lower bound.

This commit removes the lower bound set in modules.json and
the accompanying logic in init.pp.